+++ src/app/HeadlessSmoke.cpp
@@ -17,15 +17,15 @@ int RunHeadlessSmoke(const AppConfig& config, int steps) {
   core::WorldGrid world = world::BuildDemoWorld(config.world.width, config.world.height);
   core::OccupancyGridMap map(config.world.width, config.world.height);
   core::SimulatedLidar lidar(config.lidar.maxRange, config.lidar.beamCount, config.lidar.stepSize);
-  core::RobotPose pose{10.0F, 10.0F, 0.0F};
+  core::RobotPose pose{10.0, 10.0, 0.0};
 
   for (int i = 0; i < steps; ++i) {
     const auto scan = lidar.Scan(world, pose);
     map.IntegrateScan(pose, scan);
 
-    const float phase = static_cast<float>(i % 4);
-    const float vx = (phase < 2.0F) ? 0.5F : -0.5F;
-    const float vy = (phase == 1.0F || phase == 2.0F) ? 0.5F : -0.5F;
+    const double phase = static_cast<double>(i % 4);
+    const double vx = (phase < 2.0) ? 0.5 : -0.5;
+    const double vy = (phase == 1.0 || phase == 2.0) ? 0.5 : -0.5;
     const core::RobotPose candidate{
         .x = pose.x + vx,
         .y = pose.y + vy,
