+++ src/core/SimulatedLidar.cpp
@@ -5,9 +5,9 @@
 
 namespace slam::core {
 
-SimulatedLidar::SimulatedLidar(float maxRange, int beamCount, float stepSize)
+SimulatedLidar::SimulatedLidar(double maxRange, int beamCount, double stepSize)
     : maxRange_(maxRange), beamCount_(beamCount), stepSize_(stepSize) {
-  if (maxRange <= 0.0F || beamCount <= 0 || stepSize <= 0.0F) {
+  if (maxRange <= 0.0 || beamCount <= 0 || stepSize <= 0.0) {
     throw std::invalid_argument("SimulatedLidar parameters must be positive");
   }
 }
@@ -16,10 +16,10 @@ std::vector<ScanSample> SimulatedLidar::Scan(const WorldGrid& world, const Robot
   std::vector<ScanSample> samples;
   samples.reserve(static_cast<std::size_t>(beamCount_));
 
-  constexpr float kTwoPi = 6.28318530717958647692F;
+  constexpr double kTwoPi = 6.28318530717958647692;
   for (int beamIndex = 0; beamIndex < beamCount_; ++beamIndex) {
-    const float relativeAngle = (kTwoPi * static_cast<float>(beamIndex)) / static_cast<float>(beamCount_);
-    const float absoluteAngle = pose.theta + relativeAngle;
+    const double relativeAngle = (kTwoPi * static_cast<double>(beamIndex)) / static_cast<double>(beamCount_);
+    const double absoluteAngle = pose.theta + relativeAngle;
     const auto [distance, hit] = CastBeam(world, pose, absoluteAngle);
     samples.push_back(ScanSample{
         .relativeAngle = relativeAngle,
@@ -30,9 +30,9 @@ std::vector<ScanSample> SimulatedLidar::Scan(const WorldGrid& world, const Robot
   return samples;
 }
 
-std::pair<float, bool> SimulatedLidar::CastBeam(
-    const WorldGrid& world, const RobotPose& pose, float angle) const {
-  float distance = stepSize_;
+std::pair<double, bool> SimulatedLidar::CastBeam(
+    const WorldGrid& world, const RobotPose& pose, double angle) const {
+  double distance = stepSize_;
   while (distance <= maxRange_) {
     const int x = static_cast<int>(pose.x + std::cos(angle) * distance);
     const int y = static_cast<int>(pose.y + std::sin(angle) * distance);
