+++ src/core/Types.h
@@ -9,14 +9,14 @@ constexpr std::int16_t kFree = 0;
 constexpr std::int16_t kOccupied = 100;
 
 struct RobotPose {
-  float x = 0.0F;
-  float y = 0.0F;
-  float theta = 0.0F;
+  double x = 0.0;
+  double y = 0.0;
+  double theta = 0.0;
 };
 
 struct ScanSample {
-  float relativeAngle = 0.0F;
-  float distance = 0.0F;
+  double relativeAngle = 0.0;
+  double distance = 0.0;
   bool hit = false;
 };
